requirements.txt
azure-mgmt-containerinstance~=0.4.0
pylint~=1.9.1